Enhance visibility on font Fixed bug where the properties file would get overridden every time the software launches

+++ src/main/java/net/safefleet/prod/productionscheduler/data/files/PropertiesFile.java
@@ -14,20 +14,21 @@ public class PropertiesFile extends DataFile<Properties> {
 
     @Override
     protected void writeDefaults() {
-        OutputStream outputStream;
-        try {
-            outputStream = new FileOutputStream(super.file);
-            this.properties.setProperty("db.url", "localhost");
-            this.properties.setProperty("db.database", "database");
-            this.properties.setProperty("db.user", "user");
-            this.properties.setProperty("db.password", "password");
+        if(!super.file.exists()) {
+            OutputStream outputStream;
+            try {
+                outputStream = new FileOutputStream(super.file);
+                this.properties.setProperty("db.url", "localhost");
+                this.properties.setProperty("db.database", "database");
+                this.properties.setProperty("db.user", "user");
+                this.properties.setProperty("db.password", "password");
 
-            this.properties.store(outputStream, null);
-            outputStream.close();
-        } catch (IOException e) {
-            e.printStackTrace();
+                this.properties.store(outputStream, null);
+                outputStream.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
-
     }
 
     @Override
